Color palettes Colors used in the hybrid open access dashboard, based on subugoetheme::ugoe_pal()

Color palettes Colors used in the hybrid open access dashboard, based on subugoetheme::ugoe_pal()

colors_license

colors_license_unpaywall

colors_source_disclosure

Format

An object of class character of length 6.

An object of class character of length 2.

An object of class character of length 4.

Related Functions and Methods

Functions

colors_license: licenses

colors_license_unpaywall: other types of OA license information detected by Unpaywall

colors_source_disclosure: sources of disclosure of OA sponsorship